Get rid of non-portable "long double"

+++ src/applications/game_boy/states/stepping_state.h
@@ -33,10 +33,10 @@ private:
     static inline std::string s_game_window_subtitle = "Stepping";
 
     // Game loop - begin
-    static constexpr long double s_fps = 60.0L;
-    static constexpr long double s_tick_limit = 1000.0L / s_fps;
-    static constexpr long double s_cycles_per_ms = 4194.304;
-    static constexpr long double s_cycles_per_tick = s_cycles_per_ms * s_tick_limit;
+    static constexpr double s_fps = 60.0;
+    static constexpr double s_tick_limit = 1000.0 / s_fps;
+    static constexpr double s_cycles_per_ms = 4194.304;
+    static constexpr double s_cycles_per_tick = s_cycles_per_ms * s_tick_limit;
     // Game loop - end
 
     bool m_is_stepping_cycle { false };
